fix: Add request_body_size INTEGER to BIGINT migration
Large file uploads (>2GB) caused "out of range for type integer"
error when request_body_size was INTEGER in api_token_usage table.

- Change request_body_size from INTEGER to BIGINT in table definition
- Add migration script for existing installations
- Update Go model from int to int64
- Remove unnecessary int() cast in middleware

+++ api/internal/model/api_token.go
@@ -152,7 +152,7 @@ type APITokenUsage struct {
 	StatusCode      int       `json:"status_code" db:"status_code"`
 	ClientIP        string    `json:"client_ip" db:"client_ip"`
 	UserAgent       string    `json:"user_agent,omitempty" db:"user_agent"`
-	RequestBodySize int       `json:"request_body_size" db:"request_body_size"`
+	RequestBodySize int64     `json:"request_body_size" db:"request_body_size"`
 	ResponseTimeMs  int       `json:"response_time_ms" db:"response_time_ms"`
 	CreatedAt       time.Time `json:"created_at" db:"created_at"`
 }
